Refactor ebuild deduplicate logic into public API
Moved core deduplication logic from cmd/g2/ebuild_deduplicate.go to
g2.DeduplicateEbuilds to allow programmatic usage by other tools.

Resolves #393
